Move some  functions to rocketchat:ui-utils
