+++ packages/rocketchat-ui-utils/client/index.js
@@ -1,7 +1,36 @@
 import { AdminBox } from './lib/AdminBox';
 import { modal } from './lib/modal';
+import { SideNav } from './lib/SideNav';
+import { AccountBox } from './lib/AccountBox';
+import { menu } from './lib/menu';
+import { call } from './lib/callMethod';
+import { erase, hide, leave } from './lib/ChannelActions';
+import { MessageAction } from './lib/MessageAction';
+import { messageBox } from './lib/messageBox';
+import { popover } from './lib/popover';
+import { readMessage } from './lib/readMessages';
+import { RoomManager } from './lib/RoomManager';
+import { upsertMessage, RoomHistoryManager } from './lib/RoomHistoryManager';
+import { mainReady } from './lib/mainReady';
+import { renderMessageBody } from './lib/renderMessageBody';
 
 export {
 	AdminBox,
 	modal,
+	SideNav,
+	AccountBox,
+	menu,
+	call,
+	erase,
+	hide,
+	leave,
+	MessageAction,
+	messageBox,
+	popover,
+	readMessage,
+	RoomManager,
+	RoomHistoryManager,
+	mainReady,
+	renderMessageBody,
+	upsertMessage,
 };
