+++ packages/rocketchat-ui-utils/client/lib/mainReady.js
@@ -1,0 +1,3 @@
+import { ReactiveVar } from 'meteor/reactive-var';
+
+export let mainReady = new ReactiveVar(false); // eslint-disable-line
