+++ packages/rocketchat-ui-utils/package.js
@@ -10,7 +10,15 @@ Package.onUse(function(api) {
 	api.use([
 		'ecmascript',
 		'templating',
+		'kadira:flow-router',
+		'kadira:blaze-layout',
 		'rocketchat:utils',
+		'rocketchat:promises',
+		'rocketchat:notifications',
+		'rocketchat:authorization',
+		'rocketchat:streamer',
+		'rocketchat:models',
+		'rocketchat:lazy-load',
 	]);
 	api.mainModule('client/index.js', 'client');
 });
